improved template support for variables and filepaths (#35)
* improved template support for variables and filepaths

This commit fixes 2 problems with the regexp based
template parser:  variable names needed to match the
same pattern as file paths, but the matching patterns are different,
and the filepath matcher was too strict.

This commit permits filepaths which contain anything except '}'
supposing that leading and trailing spaces are not part of the path.

In addition, this commit extends the `[yaml]` argument into a
generalized entity which permits raw, yaml, or binary encoding
applied to either variables or operations like embed.  The idea
is that the value taken from a variable or operation is subsequently
_encoded_ into the document either as-is, as-yaml (including strings,
numbers, arrays, dicts,...), or as-binary (base64 standard encoding
as string, so as to map to Go json encoding of []byte).  By default,
no `[encoding]` defaults to `[raw]`.

The parsing has been restructured to go over the input yaml
and then replace strings.  Each string then is interpreted
with  `@{...}` being expanded according to variables and files
and encoding provided.

* check var syntax and remove utils.FileReader

(utils.FileReader was much more verbose than just doing it inline)