reconnect when cluster health fails for too long
Once the operator side health checks have failed for sufficiently long,
the route goes away and the sandbox watch gives connected=false

The existing logic was just sitting there, so we'd have
revtuns around arbitrarily long.

This adds a passive side health check on the cli which
reconnects when the corresponding sdxw operator status
is not connected for too long.

This fixes the symptom behind the slack bug report
```
after some time the revtuns are failing (again looks like an integration thing, the more unit level tests do not have this problem)
```

However, it is unclear why the operator health checks failed --
it is spurious behavior, not always there but somehow triggered
naturally on initial connect from time to time and more often
when re-applying a sandbox.  the tunnel was up when the health
checks failed and routing traffic correctly
if connecting directly to the proxyURL from in-cluster.