use LookPath instead of eval symlinks
In the case of the execution of `sudo signadot locald`, we are
not propagating the PATH, making any call to something not in
the sudo-provided root path not work.  This, in turn made the symlinks
not work -- not because they are symlinks, but rather because they
weren't in the sudo-provided root path.

However, sudo may disallow propagating PATH, so we use pass the
PATH-resolved abspath as the go exec.Cmd name.