correctly handle slow fork-exec in local connect (#208)
* This PR fixes a bug with slow fork-exec

Some tools, such as https://www.airlockdigital.com interfere with
fork/exec pattern in that there is an increased delay between
`cmd.Start` and when that command actually starts running.

In our fork/exec chain for `signadot local`, after `cmd.Start()`
we check if the sub-process is ready in the sense of writing its pid file.

This PR
1. increases the timeout for this from 1 to 30sec; and
2. does not return an error if there is a timeout waiting for the
   sub-process to be ready because that causes `signadot local connect`
   to report an error but not cleanup, but the background process is
   still there.  Rather it sends a warning to the log.
3. improves error reporting (depends in part on
   https://github.com/signadot/libconnect/pull/125)

* reduce error noise unless debug, fix newline in error message

* fix typo

* plumb `local connect --wait-timeout` to locald